refactor(cli): extract project command family into cli/commands/project.ts
